requirements.txt
fastapi
uvicorn
pydantic
pymongo
gunicorn
python-jose
python-multipart
passlib
PyJWT
motor
beanie
pydantic-settings
uuid
pytest
httpx
neo4j
redis
celery
fastapi-mail
stix2
pathlib
fitz
aiohttp
pymupdf
chardet